ML_deployment/Google_cloud_platform/Cloud_api/requirements.txt
Click==7.0
Flask==1.1.1
get==2019.4.13
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.1
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
scikit-learn==0.22.1
scipy==1.4.1
sklearn==0.22.1
Werkzeug==0.16.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-bigquery==1.20.0
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
pandas==0.24.2
pickleshare==0.7.5